+++ settings.php
@@ -93,6 +93,15 @@ if ($hassiteconfig) {
         2
     ));
 
+    $settings->add(new admin_setting_configselect(
+        'tool_vault/backupcompressionlevel',
+        new lang_string('settings_backupcompressionlevel', 'tool_vault'),
+        new lang_string('settings_backupcompressionlevel_desc', 'tool_vault',
+            join(', ', \tool_vault\constants::COMPRESSED_FILE_EXTENSIONS)),
+        9,
+        array_combine(range(0, 9), range(0, 9)),
+    ));
+
     // Restore settings.
     $settings->add(new admin_setting_heading(
         'tool_vault/restoresettings',